+++ crates/matrix-sdk/src/error.rs
@@ -14,7 +14,7 @@
 
 //! Error conditions.
 
-use std::{io::Error as IoError, sync::Arc};
+use std::{io::Error as IoError, sync::Arc, time::Duration};
 
 use as_variant::as_variant;
 #[cfg(feature = "qrcode")]
@@ -130,13 +130,15 @@ impl HttpError {
     pub fn as_client_api_error(&self) -> Option<&ruma::api::client::Error> {
         self.as_ruma_api_error().and_then(RumaApiError::as_client_api_error)
     }
+}
 
+// Another impl block that's formatted with rustfmt.
+impl HttpError {
     /// If `self` is a server error in the `errcode` + `error` format expected
     /// for client-API endpoints, returns the error kind (`errcode`).
     pub fn client_api_error_kind(&self) -> Option<&ErrorKind> {
-        self.as_client_api_error().and_then(|e| {
-            as_variant!(&e.body, ErrorBody::Standard { kind, .. } => kind)
-        })
+        self.as_client_api_error()
+            .and_then(|e| as_variant!(&e.body, ErrorBody::Standard { kind, .. } => kind))
     }
 
     /// Try to destructure the error into an universal interactive auth info.
@@ -153,6 +155,59 @@ impl HttpError {
     pub fn as_uiaa_response(&self) -> Option<&UiaaInfo> {
         self.as_ruma_api_error().and_then(as_variant!(RumaApiError::Uiaa))
     }
+
+    /// Returns whether an HTTP error response should be qualified as transient
+    /// or permanent.
+    pub(crate) fn retry_kind(&self) -> RetryKind {
+        use ruma::api::client::error::{ErrorBody, ErrorKind, RetryAfter};
+
+        use crate::RumaApiError;
+
+        // If it was a plain network error, it's either that we're disconnected from the
+        // internet, or that the remote is, so retry a few times.
+        if matches!(self, Self::Reqwest(..)) {
+            return RetryKind::Transient { retry_after: None };
+        }
+
+        if let Some(api_error) = self.as_ruma_api_error() {
+            let status_code = match api_error {
+                RumaApiError::ClientApi(e) => match e.body {
+                    ErrorBody::Standard {
+                        kind: ErrorKind::LimitExceeded { retry_after }, ..
+                    } => {
+                        let retry_after = retry_after.and_then(|retry_after| match retry_after {
+                            RetryAfter::Delay(d) => Some(d),
+                            RetryAfter::DateTime(_) => None,
+                        });
+                        return RetryKind::Transient { retry_after };
+                    }
+                    _ => Some(e.status_code),
+                },
+                RumaApiError::Uiaa(_) => None,
+                RumaApiError::Other(e) => Some(e.status_code),
+            };
+
+            if let Some(status_code) = status_code {
+                if status_code.is_server_error() {
+                    return RetryKind::Transient { retry_after: None };
+                }
+            }
+        }
+
+        RetryKind::Permanent
+    }
+}
+
+/// How should we behave with respect to retry behavior after an `HttpError`
+/// happened?
+pub(crate) enum RetryKind {
+    Transient {
+        // This is used only for attempts to retry, so on non-wasm32 code (in the `native` module).
+        #[cfg_attr(target_arch = "wasm32", allow(dead_code))]
+        retry_after: Option<Duration>,
+    },
+
+    Permanent,
 }
 
 /// Internal representation of errors.
