+++ crates/matrix-sdk/src/http_client/mod.rs
@@ -258,56 +258,3 @@ impl tower::Service<http_old::Request<Bytes>> for HttpClient {
         Box::pin(fut)
     }
 }
-
-pub(crate) enum RetryKind {
-    Transient {
-        err: HttpError,
-        // This is used only for attempts to retry, so on non-wasm32 code (in the `native` module).
-        #[cfg_attr(target_arch = "wasm32", allow(dead_code))]
-        retry_after: Option<Duration>,
-    },
-    Permanent(HttpError),
-}
-
-impl RetryKind {
-    pub fn error(self) -> HttpError {
-        match self {
-            RetryKind::Transient { err, .. } | RetryKind::Permanent(err) => err,
-        }
-    }
-}
-
-/// Returns whether an HTTP error response should be qualified as transient or
-/// permanent.
-///
-/// This is what decides whether to retry requests, in the [`native`] module.
-pub(crate) fn characterize_retry_kind(err: HttpError) -> RetryKind {
-    use ruma::api::client::error::{ErrorBody, ErrorKind, RetryAfter};
-
-    use crate::RumaApiError;
-
-    if let Some(api_error) = err.as_ruma_api_error() {
-        let status_code = match api_error {
-            RumaApiError::ClientApi(e) => match e.body {
-                ErrorBody::Standard { kind: ErrorKind::LimitExceeded { retry_after }, .. } => {
-                    let retry_after = retry_after.and_then(|retry_after| match retry_after {
-                        RetryAfter::Delay(d) => Some(d),
-                        RetryAfter::DateTime(_) => None,
-                    });
-                    return RetryKind::Transient { err, retry_after };
-                }
-                _ => Some(e.status_code),
-            },
-            RumaApiError::Uiaa(_) => None,
-            RumaApiError::Other(e) => Some(e.status_code),
-        };
-
-        if let Some(status_code) = status_code {
-            if status_code.is_server_error() {
-                return RetryKind::Transient { err, retry_after: None };
-            }
-        }
-    }
-
-    RetryKind::Permanent(err)
-}
